Rename build.lib to safe_git.sh.
Summary:
Yay!, this file is no longer a grab-bag of random functionality.  So I
changed the name to reflect that.

Now that the file is called safe_git, I don't need to put "safe" in
front of all the fucntions inside it, so I don't.

Get rid of the last place we still source build.lib.

That doesn't work anymore in our new world.  Luckily, it's not needed
anymore either, now that we allow safe_git.sh to be run in directories
other than workspace-root.

Test Plan:
Fingers crossed -- will try deploying after this is done, and also
keep a watch on other jenkins jobs.

(none new)

Reviewers: amos, colin

Reviewed By: amos, colin

Differential Revision: https://phabricator.khanacademy.org/D35883