Fully run svgcrush and pngcrush before the pipe.
The weekly-maintenace job was failing in a confusing way: "git status"
was saying that there were no changes pending, but then "git pull"
said it was aborting due to pending changes.  The problem turned out
to be a fairly subtle timing issue.

The old code would basically do this:
    ( pngcrush; echo "msg" ) | ( [ -n `git status` ] && commit; pull; push )

The idea is that you run pngcrush, you create the commit message
(which depends on what pngcrush does) and then you commit and push.

But in reality, things don't always go in that order.  There's nothing
that requires all the commands on the left side of a pipe to fully run
before the commands on the right; in general they run in parallel, and
only block when the stuff on the right needs some new input.

What was happening is that the git status was getting run *first*,
before `pngcrush`.  It was saying "no changes" and not committing.
Then pngcrush was running.  Then the pull was running, and it failed.

The fix is simple: do the pngcrush outside the pipe.  That enforces
the ordering we need.

Auditors: amos, colin

Test plan:
Fingers crossed