We're up to 6 worker machines now!
Tests are currently the critical path on the deploy, which we can
solve by throwing more machines at them.  Et voila!

I should maybe move this constant into vars/ so we only need to change
it in one place.  But for now `grep` works fine. :-)  (Even better
would be to get it automatically from jenkins, but I'm not sure if we
can do that.)

Auditors: dhruv

Test Plan:
Fingers crossed