Change deploy-webapp to deploy multiple branches at once.
Summary:
Right now, when two people want to combine their deploy, one of them
has to manually merge in the changes the other one makes into their
branch, and then deploy.

But now there's an easier way!  Just say
   sun: deploy csilvers + shivering-isles + the-hounds
and we'll deploy all three branches at once!

We do this by creating a new deploy branch, named after the time when
the branch was started, and we merge all those branches into it.  So
most of the groovy code is the same, it's just that we are always
deploying this new branch we created rather than the branch the user
specified.  (Note we create the new deploy-branch even in the common
case only one branch is being deployed.  That way we have a record of
every attempted deploy we've ever done.)

As a side effect of this change, we can no longer deploy branches that
start or end with whitespace (if indeed that ever worked).  Most will
consider this a feature, since deploying from phones often results in
`sun: deploy mybranch<space>`, which used to fail but will now
succeed.

Test Plan:
Fingers crossed -- I'll try it out, first on a normal deploy then on
one with plusses.

Reviewers: amos, colin

Reviewed By: amos, colin

Subscribers: benkraft

Differential Revision: https://phabricator.khanacademy.org/D36073