Fix should_deploy to "fail safe" on errors.
Sometimes should_deploy.py would crash or timeout in a way that would
cause it to give an rc of, e.g., 131.  This would cause us to do a
tools-only deploy!, since rc != 1 was interpreted as "no need to
deploy."

Now I've reversed the condition, so "rc == 0" means "no need to
deploy" and any other value means "do need to deploy."  So if
should_deploy.py fails in any way, in theory we will just end up doing
the deploy.  That could be extra work, but it's the safe way to fail.

Auditors: amos, colin

Test plan:
None