Temporarily disable building translation analytics.
The current.sqlite task has been broken for a long time.
It's currently unable to run do to lack of disk space on
the worker machine.

https://app.asana.com/0/210216962651575/336685208077069/f

The solution is to switch the type of ec2 machine being
used or attach an ebs volume, or maybe find some stuff
that can be deleted to free up space.

As a stop-gap measure I'm disabling the translation
analytics work. I believe that that will allow the
job to run with limited disk space.

Test Plan:
Fingers crossed.

Auditors: johnsullivan, james