Run these jobs at exactly 2am and 3am, not 2:xx am and 3:xx am.
Jenkins will try to load balance jobs by picking an arbitrary time
within the hour to start, if you use "H" for the minute field.  This
is recommended practice, and what we did.  The idea is you can have
"H 2 * * *" for multiple jobs, and they'll run at different times.

But we only have one job running at each timeslice (for now), and we
were running into problems because determine-test-dependencies was
starting at 3:53 -- almost an hour later than we thought!  This was
bumping into the workday on the east coast.  So I moved it up to start
at 3:00 exactly.

make-allcheck takes about 35 minutes to run now, so giving an hour of
time for it should be plenty.  (Famous last words.)

Fixes https://app.asana.com/0/31965416896056/360137809817644/f

Auditors: kevindangoor

Test Plan:
None