Send "just deployed" message only after both GAE and GCS deployed: part 1 (jenkins)
Summary:
See https://app.asana.com/0/31965416896056/361312151118327

We aborted several deploys last week when folks clicked on the link for
the new deployed version and all static files were 403ing.  It turns
out this was because we send the message saying "just deployed" after
just the GAE part of the deploy is finished.  Normally this isn't a
problem since the GAE part is almost always much faster.  However, for
this particular deploy, that wasn't the case.

This commit is the jenkins-tools half of moving this message to the
jenkins job from the python deploy scripts so that we can send it only
after both parts are deployed.  To do this, I just extracted out the
GAE and GCS deploy steps into their own parallel block that sends the
message at the end, and then ran this chunk in parallel with the tests.
For deploy-znd, it was simpler, and I just added the message to the end
of the deploy process.

Test Plan:
- verify in my editor (which has groovy syntax checking), that I don't
  see any syntax errors
- fingers crossed?  (I'll try to deploy the accompanying webapp change
  and see if it works.)

Reviewers: csilvers, nabil

Reviewed By: csilvers, nabil

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D36464