Fix up merge_from_branch() so we don't need to sync the merge-from branch.
Summary:
I don't know why this was ever considered a good idea.  (It probably
dates back to the day when the merge-from branch was always `master`.)
As it turns out, it's a terrible idea, because the merge-from branch
could be way in the past, and checking it out involves a ton of files
changing which means their timestamps change even though the files
change back to normal after we do the `git reset`.  This in turn makes
kake rebuild the world, which takes forever.

Now we just merge the branch normally, without trying to check it out
first.  To do this, we check if `origin/foo` exists and merge that if
so.

In the process, I took out some deploy-specific logic from this
routine, since it's not only useful for deploys.  As a bonus, you can
now deploy commits instead of needing to deploy branches!

Fixes https://app.asana.com/0/31965416896056/362821526345088

Test Plan: Fingers crossed

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D36493