Fix the non-default deploy messaging
Summary:

In D36464, I introduced some problems in the deploy-znd job because I
had it in my head that some functions, like `_alert` and variables like
the slack messages were globals, but were in fact just loaded in
deploy-webapp.groovy.

This commit loads in the slack messages and sets up a simple function
to message slack, taken mostly from deploy-history.groovy.  We've got a
lot of duplication cropping up surrounding slack messaging, so it's
probably worth creating a simple alertlib wrapper and having all the
deploy jobs as well as the `notify` function use it, but I've TODOed
that for now in favor of trying to get non-default deploys fixed sooner.

I also fixed an inconsistency in how we were pinging the deployer;
deploy-znd now adds the username to the front of the message, and I've
removed what was for deploy-webapp an extra ping from the "just
deployed" message itself.

Test Plan:
- try a non-default deploy from jenkins, fingers crossed!

Auditors: csilver, nabil