Don't decrypt secrets when it's not necessary.
Summary:
We're seeing a steady trickle of "secrets.py is newer than
secrets.py.cast5."  I //think// this happens when multiple jobs
decrypt at the same time.

One way to help with that is to not decrypt as often.
secrets.py.cast5 changes very rarely, so we shouldn't be decrypting
all the time.  We were before, but after this change we won't be.

A real locking solution would be better, or perhaps just having the
secrets-dir be per-job rather than global.  That's a TODO for now.

Test Plan: Fingers crossed

Reviewers: amos, colin

Differential Revision: https://phabricator.khanacademy.org/D36514