Make sendToSlack() more robust to errors in onMaster().
Summary:
`notify.sendToSlack()` is the way we say there are errors (or
success!), but if that function call itself throws an error, then we
don't see anything on slack.  So we want to make this function as
robust to errors as we can.

I added a fallback in case `onMaster()` fails.  `onMaster()` does a
bunch of stuff; of particular concern it tries to pull from github.
If that fails we just proceed along.  In the worst case we might get
an old version of alertlib, but in the (very) common case the pull
wouldn't have added any value anyway.

Fixes https://app.asana.com/0/25429045350268/359010280811280

Test Plan: groovy vars/notify.groovy

Reviewers: amos, colin

Reviewed By: amos, colin

Subscribers: nabil

Differential Revision: https://phabricator.khanacademy.org/D36945