Simplify how we handle secrets on Jenkins
Previously, we always kept secrets on Jenkins in a special special directory,
and then added it to the python path.  We did this to make sure we never had
secrets when we didn't want to.  But this doesn't play nice with wanting
secrets to be in `shared/`.  So now we just do a more normal thing, and decrypt
within webapp.  With groovy, it's plenty easy to make sure we clean up on exit.

This actually simplifies the whole assembly a good bit.  It also fixes a bug;
in the past we would fail silently if decrypting secrets didn't work right.
Now we'll error as intended (and still clean up after ourselves!).

Test plan: deploy it, see end to end tests report to slack

Auditors: csilvers