Change how jenkins handles secrets, and update for the new location of secrets
Summary:
This commit makes several changes, all in support of the new location of
secrets in webapp in D36597, which moves it to the shared directory.  The first
change is simply to update paths as appropriate.

In addition, previously, we always kept secrets on Jenkins in a special special
directory, and then added it to the python path.  We did this to make sure we
never had secrets when we didn't want to.  But this doesn't play nice with
wanting secrets to be in `shared/`.  So now we just do a more normal thing, and
decrypt within webapp.  With groovy, it's plenty easy to make sure we clean up
on exit.

This actually simplifies the whole assembly a good bit.  It also fixes a bug;
in the past we would fail silently if decrypting secrets didn't work right.
Now we'll error as intended (and still clean up after ourselves!).  Finally,
this requires substating alertlib to include D36956, and setting the
appropriate environment variable.

This reverts 1da50683 and ae1c168d with some additional changes, unreverting or
clobbering 60d1de01, 7ac69aac, f7741fd0, and 46f5bf0e.

Test Plan: Deploy it, along with D36597, and see slack messages work.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D36959