Add backwards-compatibility for older webapp
I was hoping we could avoid doing this, by just deploying the secrets move at
the same time in webapp and jenkins-tools.  But jenkins-tools needs to be able
to do a little bit of work before even checking out the new webapp!  So I had
to add some compatibility.  This will also hopefully save a revert if I mess
things up.

Test plan: deploy with this, not have it fail

Auditors: csilvers