Prompts while waiting for input (for set-default and finish), after a while.
Summary:
One problem is that people can forget its their turn to set-default,
or to finish.  Then we time out after an hour and they're like 'oops'.
Now we remind them periodically.

The implementation here was annoyingly complex.  I wanted to have two
threads running at the same time, one sleeping-and-warning, and the
other waiting for input, and then we continue on when either of them
finishes.  But I don't think it's possible for one thread to interrupt
another's `sleep()`, so it would have had to be busy-sleeping
(sleeping 1 second at a time) which didn't seem good.

So I went another approach where I split the input request into 4
input requests, each with their own timeout.  We just do 'input; warn;
input; warn; input; abort'.  This means that if someone says `sun: set
default` in the instant we're warning (between the inputs) we won't
properly identify it, but a) we'll give a good error message so they
can retry; and b) that's not going to happen anyway because everyone
is good and set-defaults within a few minutes of being prompted.
Right?

Addresses (to my mind fixes)
https://app.asana.com/0/25429045350268/364810432654224

Test Plan:
groovy jobs/deploy-webapp.groovy
groovy jobs/deploy-webapp_slackmsgs.groovy

But mostly fingers crossed.  We'll see on the next deploy.

Reviewers: amos, colin

Reviewed By: amos

Subscribers: benkraft

Differential Revision: https://phabricator.khanacademy.org/D36999