Move the `return` outside the `withTimeout`.
I'm not sure what's going on, but I wasn't seeing any returning going
on -- we kept looping even when the user clicked 'ok'.  I'm guessing
that the `return` was somehow returning from the block inside the
`withTimeout` and not from the function entirely.  That makes sense
since the block inside `withTimeout` is actually a closure.

Auditors: amos, colin

Test plan:
Fingers crossed