Move emergency-rollback to groovy.
Now all jobs are on groovy.  I wanted to keep emergency-rollback on
not-groovy, thinking it would be more stable, but the opposite has
turned out to be true: as the one outlier, it would break when we
refactored stuff and forgot to test on it.

Auditors: benkraft

Test Plan:
Will try rolling back after this is live.