Don't trust git when deciding whether to roll back
Summary:
In the case where we're somewhere in set-default, and we haven't yet finished
with success, we shouldn't trust git as to whether we've rolled forward yet --
it definitely doesn't know!  If we can't get it from the API, we should try a
rollback anyway; we already know what the good and bad version s should be.
Depends on D37902 in webapp to add the flag I used.

Test Plan: cross fingers

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D37903