Update check_live_workers cronjob to send alerts regardless of error state
Summary: In order to automatically close alerts in the Alerta dashboard, cronjobs must be setup to ping alerta when jenkins workers are back to normal after an error state. This update makes check_live_workers send an alert with the same args (save severity, which gets translated to ‘cleared’ in the alerta script) to resolve the issue on the dashboard.

Test Plan: Tested locally

Reviewers: nabil, kevindangoor

Reviewed By: nabil

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D38234