Do more checks on test-splits before using them
If test-splitting does something weird, we currently just press on -- and if
there are older test-split files floating around in the workspace, we may
accidentally use them.  (In the case I investigated, "something weird" is a
`tests_for` failure -- which doesn't bubble up since we don't use
`-o pipefail`.  So we just get empty output.)  Now we'll blow away any old
ones, and also check that we got the number we thought we should.

Test plan: crossed fingers

Auditors: csilvers