Handle the fact secrets have moved up a level in the directory tree.
I was installing some secrets in the workspace of the job that needed
them, which is great.  But groovy has a habit of destroying and
recreating the workspace at a whim.  So I put them in a new place, a
directory higher.  Now the scripts look in the new location.

Auditors: dhruv

Test Plan:
Fingers crossed, will try a react-render-server deploy.