Always clean .pyc files before doing any work in webapp.
A refactor can cause a .py file to go away, but the .pyc file to stick
around.  We'd like to make sure that .pyc file doesn't cause any
trouble, so we just nuke it.

We could do a better check to make sure the .pyc file //should// be
deleted, something like
   find . -name '*.pyc' -exec sh -c 'test ! -e "${1%c}"' -- {} \; -print0
but it's probably slower to do all those bash invocations than it is
just to regenerate all the .pyc files at need as the script runs.

Auditors: benkraft

Test Plan:
Fingers crossed -- we'll see script failures if something is wrong.