Add mypy and unit tests to the khanalytics-test jenkins job
Summary:
I'd previously set up a jenkins job for "testing" khanalytics, but it
wasn't too useful because all it did was run the linter and flow. We
weren't running tests or mypy because I needed to figure out how to get
python3 set up with a virualenv on jenkins.

This commit sets up a python3 virutalenv so that we can run the tests.
To do this, I stole some of the code from withVirtualenv.groovy,
modifying it accordingly so that we can have both the python3 and
python2 virtualenvs side-by-side.

Test Plan:
- push to a temporary testing branch, and point the khanalytics-test
  job on jenkins at that branch
- go to https://jenkins.khanacademy.org/job/deploy/job/khanalytics-test/
- select "build with parameters" (and use #bot-testing as the room)
- start the job
- verify the job completes successfully
- verify I see a note that the tests passed in the logs
- verify that I see that it ran mypy in the logs and there were no
  errors.

Reviewers: ragini, gustavosilva, csilvers

Reviewed By: ragini

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D38826