`using-test-workers` is now a lock-label, not a lock-name.
The difference between lock-labels and lock-names is that a label can
apply to more than one resource, while a name applies to just one.  In
our case, we want to use labels because we want to support several
banks of test-workers, and let the lock acquire just one of those
banks.  So we have a name for each bank but they all share a label,
and when you want to reserve a bank it just picks one.

It's a little confusing, and I wish that they had organized their
locks a different way, but c'est la vie.

This is the code follow-up to a config change I made at
   https://jenkins.khanacademy.org/configure
where I went to the "Lockable Resources" section and replaced
    Name: using-test-workers
    Labels: <empty>
with
    Name: test-workers-1
    Labels: using-test-workers
    <added a new lockable resource>
    Name: test-workers-2
    Labels: using-test-workers
and then went to the "Amazon EC2" section and clicked on "advanced"
and raised the instance cap to 12.

Auditors: amos, colin

Test Plan:
`groovy jobs/webapp-test.groovy` without errors.  But I'm going to try
to run a deploy and test at the same time and see what happens!