Allocate an executor in notify(), that is used for the duration of every job.
Summary:
When I first wrote all these groovy scripts, I did so with the idea of
using an executor (a "thread" in jenkins that can do work) as little
as possible.  Many groovy commands don't require an executor, so you
could do something like:
    onMaster() { foo }
    prompt ...
    onMaster() { bar }
and open up an executor for other purposes while waiting for the user
to respond to the prompt.

The problem with the above is there's no guarantee that the two
`onMaster` calls will use the same workspace directory; one might use
   jobs/workspace
while the other might use
   jobs/workspace@2
(This is not a problem if only one instance of a particular job can run
at a time, but that's not always the case.)  The only way to ensure
that a job uses the same workspace throughout is to allocate an
executor at the beginning of the job and keep it all the way through
until the end.

So that's what we do now.  It's not quite as efficient in terms of
executors, but much safer.

The old `onMaster` is now deprecated, since it happens automatically
in `notify`.  (Basically, we're //always// on master now.)  The only
functionality that's useful to keep from the old `onMaster` calls is the
timeout, as a bound for how long individual steps should take.  (I
also gave each job a global timeout that's set in notify.)  So I
changed all the existing `onMaster` calls to `withTimeout`.

Test Plan:
Fingers way crossed.  I'll deploy this and then try running a few
jobs, like `webapp-test`.

Auditors: amos, colin