Fix up places that wanted to be called outside a node.
Now that we enter a node (via `onMaster`) at the beginning of each job
and don't leave until the end, it's not appropriate to have internal
functions that expect to be called outside a node (and allocate their
own).  So fix that up.

It simplifies some things; hopefully it doesn't break anything else.

Auditors: amos, colin

Test Plan:
Fingers crossed