Be more cautious when trusting the environment.
Now that we have a master node running all the time, when test-worker
nodes spawn they inherit the environment from the current master node.
So it no longer works to check whether stuff is set on `env` as a way of
caching whether we've done work in this node.  I have to fix them all
to at least be more careful.

Auditors: amos, colin

Test plan:
Fingers crossed