Be more robust against /tmp/make_check.run not existing.
Lots of things touch and delete this file, and it's possible for some
of them to be running at the same time now that we have two banks of
test-workers.

We actually totally need to revisit how this monitoring check works in
a world where we have multiple test-worker banks -- and other types of
worker machines as well -- so I've disabled it for now.

Auditors: amos, colin

Test plan:
Fingers crossed