Do our best to install alertlib deps before using alertlib.
We (somewhat hackily) install all the alertlib deps as part of
webapp's `make install`, so usually when we want to send to alertlib
from a jenkins job, the deps have been set up to do so.

But occassionally this is not the case.  For instance, webapp-deploy
can fail before `make deps` is run, for instance if there's a merge
conflict.  It will then try to send to alertlib but potentially fail
because the deps aren't installed.  This is exactly what happened in
    https://jenkins.khanacademy.org/job/deploy/job/deploy-webapp/1612
which was the first deploy after the workspace had been deleted.

By installing deps before trying to use alertlib, we should fix this
problem.

Auditors: amos, colin, amy

Test Plan:
I'll try to deploy branch `gtp`.  It should fail with a 'need to
merge' error this time, instead of silently.