Change webapp-test to be able to run tests on several (merged) branches.
Summary:
This will be useful for testing a branch when you're next in line to
deploy.  You can do `sun test: currently-deploying-branch + mybranch`
and it will test what your deploy will be like, assuming the current
deploy succeeds.

While in the area, I got rid of a call to `make clean` that was happening
twice: in a function that someone called, and then right after that function
returned.

Test Plan: Will try it in real life!

Reviewers: amos, colin

Reviewed By: amos

Differential Revision: https://phabricator.khanacademy.org/D39396