Use the normal `notify` after all.
I thought we didn't need our own executor, but we do because `notify`
does stuff when tests pass (or fail) and it needs an executor to do
that.

This means that we take up an extra executor while the sub-tasks are
running, that is just sitting there doing nothing but blocking up the
queue, which is not ideal.  But make-allcheck runs early in the
morning when not a lot else is going on (I think) so it's not too
great a cost.

Test plan:
Will see how the job runs tomorrow!

Auditors: jacqueline, benkraft