Jenkins job for deploying khanalytics
Summary:
See https://khanacademy.atlassian.net/browse/INFRA-24

Right now the deploy process for khanalytics happens manually on
developers' machines. This requires some manual setup, and doesn't do
anything that requires tests/lint/type checking to be passing on the
commit being deployed.

This commit creates a jenkins job to deploy khanalytics. All the
deployment scripts and whatnot exist in the khanalytics-private
repository, which contains the main khanalytics codebase as a
submodule. To deploy, we therefore substate khanalytics into
khanalytics-private, run tests on what we just substated, and then if
tests pass, deploy to production.

This was fairly straightforward; the only non-obvious change here was
that I needed to make it possible to call
`update_submodule_pointer_to_master` on the `safe_git.sh` command line.
While it was previously a non-underscored (public) method in that file,
there was no way to call it because that function expected to be called
from inside a repository, and the script itself enforces that you're
calling it from outside a repository. I changed the convention of that
function to be like the rest of the functions in safe_git.sh -- it now
expects to be called from outside the repository and passed the
directory of the respository as an argument.

There's a lot of potential TODOs here, including: customizing what
commit we're deploying instead of assuming master, not assuming that
we're the only thing talking to kubernetes on jenkins, better slack
messaging, running integration tests, maybe not pushing the substate
commit if tests don't pass, etc., etc. But this is already a big step
up in both developer experience and reliability, so I wanted to get it
out there.

Test Plan:
- push this to a WIP branch and point jenkins at this WIP branch for
  the khanalytics-deploy job
- run the khanalytics-deploy job
- verify it succeeds
- verify it created a substate commit in the khanalytics-private repo
- verify it ran the khanalytics-test job
- use `kubectl get pods` to see that the age of the deployed pods is
  consistent with it having deployed a new version

Reviewers: csilvers, amos, ragini

Reviewed By: csilvers, ragini

Differential Revision: https://phabricator.khanacademy.org/D39599