Be sure to `git pull` in the khanalytics submodule
Summary:

In a previous commit, I created a jenkins job for khanalytics that
would automatically substate the khanalytics submodule in the
khanalytics-private repository. This worked fine when I tested it the
first time because we were doing a fresh clone. However, when trying to
deploy again, I realized we weren't actually making a substate commit.
This is because we're updating the submodule pointer to master, but we
weren't actually pulling in upstream changes in the submodule.

This commit adds a pull step in the khanalytics submodule so that we
actually have changes to substate.

Test plan:
- Push to a temporary WIP branch, run the job, verify that it now
  creates a substate commit, where before it was claiming there were no
  changes.

Auditors: csilvers, ragini