Put the alert inside the try/finally for rollbacks.
That way, even if the alerting-about-rollback fails, we still get some
notice.

I could have a separate try/finally for the alerting, and try to roll
back even if we can't alert, but I figure this way is cleaner, and
this "shouldn't" ever happen (though it did in
https://jenkins.khanacademy.org/job/deploy/job/deploy-webapp/1884/console),
so I don't feel the need to get all fancy on it.

Addresses https://app.asana.com/0/31965416896056/440870381815874/f

Auditors: benkraft

Test plan:
None.