Do some echoing every time we throw.
We often catch and rethrow, doing some work before rethrowing.  But
it's very possible that that extra work might itself throw, which
means the original error would be swallowed (since we never get to the
rethrow in that case).

So now in all our catch-and-rethrow cases I log the error before doing
the extra work, so at least there's some indication of what went wrong
in the logs.

Auditors: benkraft

Test plan:
groovy jobs/deploy-webapp.groovy