Centralize the default number of test-workers we use per job in a single file.
Summary:
Originally, I had planned that every job could decide how many
test-workers to use, and you could adjust this based on how fast you
needed a particular job to run.

And maybe one day we will get to that goal, but for now we can't
because the lockable-resource plugin doesn't allow us to lock machines
individually; we need to lock an entire "bank" at once.  So every job
might as well use the same number of workers as every other job.

To make it easiser to adjust this number, I made it a constant.  I
also added instructions for how to update the number correctly (you
have to coordinate with jenkins).

Test Plan: Will try running webapp-test after this is deployed.

Reviewers: amos, colin

Reviewed By: amos, colin

Differential Revision: https://phabricator.khanacademy.org/D39860