Add some tests for safe_git.sh.
Summary:
There's still a lot more that could be done here, but this is a start.

The goal is to be able to simplify the safe_git routines to do a bunch
less work.  But we want to make sure that they still work correctly!
These tests try to exercise the corner cases that we've built up
safe_git.sh to handle over the years, particularly when it comes to
submodules changing in strange and exciting ways.

Some changes to make safe_git.sh work under OS X.

This isn't a complete list since I don't have an OS X box to test on,
but it's at least a help:

1) Use gtimeout instead of timeout when timeout isn't installed
2) Don't try to use the `-f` flag to readlink.

Fix the submodule-pointer test, which wasn't updating the submodule properly.

Test Plan:
./safe_git_test.sh -l
./safe_git_test.sh test_safe_sync_to_origin__local_changes
./safe_git_test.sh

./safe_git_test.sh

(same)

Reviewers: benkraft, jacqueline

Reviewed By: benkraft

Subscribers: jacqueline

Differential Revision: https://phabricator.khanacademy.org/D40155