Add instructions for running safe_git_test.sh on OSX
Summary:
`safe_git.sh` and `safe_git_test.sh` are of course designed to run on
linux, since that's what jenkins is running, but unfortunately there
are some parts that don't work out of the box on OSX, which is annoying
for running the tests.

I don't think we want to add OS-specific functionality here, since this
really does target a single platform, so this commit adds some extra
things you may need to do on OSX to get the test script running.

Test Plan:
- run the commands listed in the comments
- `./safe_git_test.sh`
- verify that the tests run succesfully

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: jacqueline

Differential Revision: https://phabricator.khanacademy.org/D40253