Make it more obvious when the `safe_git_test.sh` tests fail
Summary:
I was trying to add a new test to the `safe_git_test.sh` file and was
very confused that it wasn't being run. It turns out this is because
the tests were already failing earlier, but I didn't realize that
because I didn't know to look for the "PASS" echoed at the end of the
file.

This commit makes the `_verify_at_master` assertion function echo a red
`FAILED` after the diff if the assertion fails, to make it more obvious
that a test actually failed.

Test Plan:
- `./safe_git_test.sh`
- verify that now I see a "FAILED" when a test fails (and that I
  therefore realize that these tests are currently failing)

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40261