Fix failing safe_git test
Summary:
The safe git tests were failing at the "Change what a submodule points
to" step. This resulted from a problem pulling in changes after
changing what the submodule points to in origin, which means that we
never actually committed the change to `.gitmodules`. (The error during
`git pull` was `fatal: refusing to merge unrelated histories` because
`pull` was trying to merge the new submodule's history into the old
submodule's history that it had checked out.

This solution here is just to `--reset hard` to `origin/master` instead
of pulling.

I also fixed an issue where `sed -i` was interpreting `-e` as the backup
prefix for in-place modification and leaving around an untracked file
`.gitmodules-e` in the origin repository, which made the diff come up
as nonempty.

Test Plan:
- `./safe_git_test.sh`
- verify that it now passes!

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40271