Failing testcase for submodule rollback
Summary:
See https://app.asana.com/0/31965416896056/443069942927568/f

We've seen two cases (one in webapp, one in khanalytics), where jenkins
has failed to do the correct thing for submodules. It's not clear
whether or not they are the same issue. This commit adds a testcase for
an interesting bug that happens when we try to do the following:
- on origin:
    - make a commit in `subrepo1` (which is included as a submodule in
      `repo` one; but in the toplevel repository for `subrepo1`)
    - substate that new commit into `repo`
- sync to the local copy using `sync_to_origin`
- roll back the substate commit on origin (using `reset --hard HEAD^`)
- sync to the local copy using `sync_to_origin`

When this happens, the local copy ends up on a subrepo commit ends up
on a commit that is neither the one we were rolling back from nor the
one we were rolling back to. Mysterious!

I don't really know what is going on here; I want to commit this
testcase just in case I don't figure this out before my support
rotation ends, to make handoff easier.

Test Plan:
- `./safe_git_test.sh`
- verify the tests fail

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D40321