Do the timeout before the flock, not after.
Both timeout and flock are commands that take the actual command to
run as an argument.  So when using both, you can put them in any
order.

But in our case, `timeout` might be a shell alias sometimes.  In that
case, it has to go first so the shell can handle it properly.  We do
that.

This should get safe_git.sh working better on OS X without the need to
do anything special to install `timeout`.

Auditors: colin, jacqueline

Test plan:
Fingers crossed.  I did run
    timeout 120m flock -w 7230 /var/tmp/flock ls
and it worked fine.