Use `;` instead of `&&` in most places.
When running in `set -e` mode, `;` is better than `&&` since the
script will crash whenever a command fails if it ends with `;`, but
not necessarily if it ends with `&&` (I think).  It's clearer in any
case.

Inside backticks errors aren't fatal even in `-e` mode, so I left
those `&&`s as they were.

Auditors: colin, jacqueline

Test plan:
./safe_git_test.sh