Add some more tests, particularly tests for nested submodules.
I realized we weren't testing submodules that had submodules of their
own, that is that we use `--recursive` properly everywhere.  So I
added some tests for that.

I also added a few more tests for how we might munge the local copy of
the repo, to make sure that our cleaning is correct.

Auditors: colin, jacqueline

Test plan:
./safe_git_test.sh