Do introspection to figure out the test names!
This should avoid the trap I fell into before of adding a test but
forgetting to add it to ALL_TESTS, so it was never run.

Auditors: jacqueline, colin

Test plan:
Ran
   ./safe_git_test.sh -l
and saw the same list of test before and after this change.

Then I added a line
   test_add_a_file() {}
to the end of the file and ran `./safe_git_test.sh` and it said:
   Defined multiple tests with the same name:
   test_add_a_file