Revert "Speed up safe-git's destructive-checkout."
This reverts commit 8bc6fa115157d5acc30e52c13a221a6efcae057b.

There's a change here that breaks the current implementation of update_submodule_pointer_to_master.

While it looks like the old removed code intends to update submodules to the root's commit, it does not in fact do that. Here the new "--recurse-submodules" actually updates the submodules.

This makes i18n-download-translations, update-i18n-lite-videos, and i18n-update-strings silently fail to update the subrepo. I think this might also break substateKhanalytics.

I'm going to revert to unblock our translators this because I don't really know what the correct fix is here. As best I can tell, the code that updates submodules does not expect pulling to be destructive with respect to submodules, but this code looks like it intended to be so.
If I had a better idea of which side should change and was convinced I could do it without breaking something else, I would update it, but it's not totally clear to me. I'm happy to give it a spin if you would like me to.

Auditors: csilvers, colin, jacqueline