Don't use `git commit -a` by default anymore.
We are having trouble where tests_for and other jenkins jobs are
adding back in submodules that had been deleted.  It actually adds
them back in as nested repositories, not submodules, which means
that the dirs are around, complete with a .git file, and when we do
`git add -a` it adds them back in but in a harmful way.

The underlying cause of the problem -- why was pyjwt still around even
after a git clean -- is `.gitignore`.  We have `.pyc` in our
.gitignore, which means that `git clean -ffd` does not fully clean
these directories.  Then, at push time we do a `git commit -a` and
those dirs get added back in (though with no content, just a `.git`
file).

There are various ways to solve this problem, but the most effective
is to not use `git commit -a`, which is dangerous with submodules.
Instead, I explicitly list the files I want to check in.  For some
scripts that only work in intl/translations (or another repo without
submodules), I kept the `-a`, though explicit this time.  That should
be safe enough.

See commit f2125128e9c766f712450a0993a4a7c73ab3ad5e, generated by
  https://jenkins.khanacademy.org/job/misc/job/determine-test-dependencies/282/
for an example of where things went wrong, though it's hard to tell
from the console log here.

Another solution might be to use the `-x` flag with `git clean -ffd`,
though that's scary because sometimes we *want* to keep the
.gitignore'd files, e.g. secrets installed at ec2-install time.

Auditors: joshuan

Test Plan:
Fingers crossed, we'll have to see how these jobs go.